+++ R/scienceverse-package.R
@@ -1,0 +1,7 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom faux nested_list
+## usethis namespace: end
+NULL